requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
yfinance
pyyaml
